requirements.txt
torch
torchaudio
soundfile
hydra-core
plotly
h5py
wandb
einops
pyroomacoustics
librosa
tqdm
pandas
pesq
pystoi
onnxruntime
matplotlib
torchcde
ninja